+++ controllers/tests/pick_up.py
@@ -1,3 +1,4 @@
+from typing import Dict, List
 from tdw.tdw_utils import TDWUtils
 from magnebot import TestController, Arm
 from magnebot.action_status import ActionStatus
@@ -14,14 +15,13 @@ class PickUp(TestController):
         self.target_object_1: int = -1
         self.box: int = -1
 
-    def init_scene(self, scene: str = None, layout: int = None, room: int = None) -> ActionStatus:
+    def _get_scene_init_commands(self, magnebot_position: Dict[str, float] = None) -> List[dict]:
         # Add some objects to an empty room. Record their object IDs.
         self.target_object_0 = self._add_object("jug05", position={"x": -0.408, "y": 0, "z": 0.428},)
         self.target_object_1 = self._add_object("jug05", position={"x": -1.76, "y": 0, "z": -1.08})
         self.box = self._add_object("basket_18inx18inx12iin", position={"x": 0.03, "y": 0, "z": -2.38},
                                     scale={"x": 1, "y": 0.5, "z": 1})
-
-        return super().init_scene()
+        return super()._get_scene_init_commands(magnebot_position=magnebot_position)
 
 
 if __name__ == "__main__":